backend/requirements.txt
fastapi==0.110.0
uvicorn==0.31.0
mysql-connector-python==8.3.0
pydantic==2.7.0
python-multipart==0.0.9
python-dotenv==1.0.1
